Added game submit route to app.js
Created gameSubmitRoutes and submitController files

	modified:   app.js
	new file:   controllers/submitController.js
	new file:   routes/gameSubmitRoutes.js
